+++ packages/preset-umi/src/features/icons/esbuildPlugins/esbuildAliasPlugin.ts
