+++ packages/preset-umi/src/features/icons/esbuildPlugins/esbuildExternalPlugin.ts
