+++ packages/preset-umi/src/features/icons/extract.ts
